feat(timetable): integrate weather banner and enhance UI components
- Added WeatherBanner widget to display current weather information in the timetable tab.
- Replaced TimetableStatusBanner with WeatherBanner for improved user experience.
- Updated TimetableAppointmentCard and TimetableAppointmentCupertino to have a consistent border radius of 16.
- Refactored TimetableControlPanel to manage term start date input more effectively with a TextEditingController.
- Enhanced error handling in TimetableController during data synchronization processes.
- Updated main.dart to include global error handling and improved initialization of Rust FFI bridge.
- Added geolocator dependency for location services and updated related platform files for macOS and Windows.
- Removed flutter_native_splash dependency and its configurations from pubspec.yaml.
- Updated various dependencies in pubspec.lock and Cargo.toml for better performance and stability.

+++ lib/core/settings/presentation/settings_providers.dart
@@ -74,7 +74,7 @@ class SettingsController {
   }
 
   void _syncWithRust() {
-    crawler.updateProxyConfig(port: _state.value.proxyPort.toInt());
+    crawler.updateProxyConfig(port: _state.value.proxyPort);
 
     if (isWeb) return;
